backup_petorlandia/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.16.2
APScheduler==3.11.0
attrs==25.3.0
blinker==1.9.0
boto3==1.38.44
botocore==1.38.44
cachelib==0.13.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Babel==4.0.0
frozenlist==1.7.0
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
mercadopago==2.3.0
msgspec==0.19.0
multidict==6.5.1
packaging==25.0
pillow==10.4.0
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.9
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
qrcode==8.1
requests==2.32.3
s3transfer==0.13.0
six==1.17.0
SQLAlchemy==2.0.41
twilio==9.5.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.1.2
yarl==1.20.1
